Introduced initial elements of a Timeline
Created the skeleton of a timeline using boxes & containers (and a lot of help from an online tutorial!). Needs to be refined and i need to make sure i fully understand all the code and how it interacts. Also, would be nice to move some of the classes out of the main code (and reference/export it in, or #include if that is possible in HTML) as its now getting very long and more difficult to read.