Updated values from Absolute to Relative
Updated a lot of the text, image and object measurements from pixels to % or VW to allow the page to scale down on smaller screens. Additionally added some flex boxes at the bottom of the page to replace the UL text to make the site more interesting to look at.