Image carousel added to the page
Adding JS & CSS files for the image carousel - should be working (had a bug for a long time that was driven by position of the <script> too early in the HTML, meaning variables used in JS wernt yet calculated when called upon in the carousel) - all fixed now and should be working. Image files also changed to local so need to upload.