Built in Autoplay & DOM manip on button
Built in auto play functionality to Carousel via a recursive function and introduced direct DOM manipulation on the button (changing properties via JS) so that the text now changes alongside its colour when clicked. had issues with turning autoplay on and off but now fixed. (feeds in true/false to function when button clicked to return out of recursion if needed). Also fixed error on timeline alignment (was due top margins not set to 0% 0% for the "right" class)