Add Collapsible boxes to time line
Added collapsible boxes to time line with CSS & JS that hides the timeline text and then unhides it when you click the title - wanted to do a transition using box height but was too complicated! so sticking to this method for now. Timeline still WIP as issues with overlapping layers.