Added circles to Carousel, plus other site updates
Added circles to the bottom of the Carousel to allow user to see current position - at the moment they only change colour on movement when clicking the arrows/autoplay (by affixing new class based on current index), next step is to make the circles clickable to change position in carousel. Other changes include content tweaks/typo fixes and animation refinements so that the typing effect works on all size screens.